fixes bugs in peeking iterator
PeekingIterator was not handling null properly and was not following the
iterator contract and throwing NoSuchElementException when it should.
The null bug was if the source iterator returned null for next() this
would cause peeking iterator to falsely claim there was no next.  When
the iterator did not have a next it would return null for next() instead
of throwing an exception.

Fixed both bugs and added unit test.  Looked at the test coverage of the
class and added a few other test that cover more then these bugs to get
100% coverage.