Add more test cases to bulk skip test
Adds tablet gaps of size 1,2,3,...,9 tablets to the bulk import load
plan in the skip test. The goal of this is to attempt to cover different
edge case w/ the bulk import skip property.  Like when the skip property
is set to 4 the load plan will have gaps of 3,4,5 tablets that is has to
deal with.