Fixed testRootTabletUpdate test method to re-host root tablet
The testRootTabletUpdate test was setting an operationId on
the root tablet, which ends up causing the tablet to become
unhosted causing the remaining tests to time out. Added code
to the test to unset the operationId and set a future location
for the root tablet so that it will once again be hosted.

Closes #5440