Fixes race condition in deleting namespaces
Namespace deletion checked that no tables existed in a namespace in the
Accumulo client.  No checks were done on the servers side.  This has
race conditions where the set of tables in a namespace changes after the
client side check.

Added a server side check that is done after the namespace lock is
acquired.  The client side check should be removed, however this would
not be safe to do in a bug fix release.  If the client side check is
removed in 2.1.4 and a 2.1.4 client runs with a 2.1.3 server then
nothing would do the check.

To test this change the client side check was commented out and
NamespaceIT.deleteNonEmptyNamespace() was run.  This shows the change
works now, but regressions in this new code will not be detected.

In main/4.0 the client side check can be removed.