fixes accumulo-cluster bug with starting multiple compactors
The parse_config function was dynamically creating variables that were
scoped to the function. When the code attempted to read these variables
in the execute_command command function it did not find them. This
caused the script to only start one compactor per host even if multiple
were configured. Added -g to declare builtin to make the variables
global.