Modified TabletMetadata to expose loaded column timestamp
Modifed TabletMetadata.getLoaded to return a list of
Triple<StoredTabletFile, FateId, Long> instead of a
Map<StoredTabletFile, FateId>. The Long in the triple
is the timestamp from the Key, and is used in the
TabletManagementIterator to return the TabletMetadata
to the TabletGroupWatcher if the column for the Fate
bulk import transaction is over 24 hours old. In the
TabletGroupWatcher the Manager logs this at a WARN
level, and could increment a metric, to alert users
to the fact that they have a failing or stuck bulk
import.

Closes #5175