Added 'accumulo upgrade --prepare' command to prepare for upgrade (#5438)
Added `accumulo upgrade --prepare` command which is intended to be used
after an instance is shutdown in preparation for an upgrade to the next
minor or major release. This is not intended to be used for bugfix
releases. `accumulo upgrade --prepare` will validate that no Fate transactions
exist, create a marker node in ZooKeeper that this utility has been run, then
proceed to remove all locks in ZooKeeper for the Accumulo servers.

Accumulo server processes will not start with the upgrade marker node present in
ZooKeeper, so if this utility is run by mistake on a bugfix upgrade, then the
user will need to remove it manually. If the user decides to abort the upgrade
process after running this utility, then they will need to remove the marker
node from ZooKeeper manually as well.