WIP implements removing locks in zoozap
Modified the accumulo-cluster script to find all of the host and ports
for all of its pid files accross the cluster.  This is the complete set
of hosts that accumulo cluster script knows about. Added an option to
ZooZap to exclude host and ports and passed the set of derived from pid
files.  This will cause the prune command to remove any process in
zookeeper that does not have an associated pid file.  Processes w/o a
pid file are unknown to the accumulo-cluster script.

Modifed the accumulo-service list command to include pid and port, this
seems like a generally useful change.  This was needed so the
accumulo-cluster script can determine the ports associated w/ pid files
on a host.