Added 'a' short option to Shell merge --all command
Prior to #5450 the all option in the merge command
used an empty string for the short option. As part of
which caused a failure because commons-cli no longer
allows an empty string as the short option (see [1]).
without a short option, which is allowed, but that
appears to have caused failures in ShellIT when
the `merge --all` command is used. This commit
simply adds 'a' as the short option and removes
the Option.Builder modification.

[1] https://github.com/apache/commons-cli/commit/0ece5e45c0d5c713f1c3888281f8eb32c252d5db