Fixed MergeCommand to use longOpt for all option
Prior to #5450 the all option in the merge command
used an empty string for the short option. In 5450
we changed MergeCommand to use the Option.Builder,
which allows specifying a long option only, but
we didn't use the long option when getting the
option value which causes a ShellIT failure.