Changed log4j2 statusLogger level from info to warn
After bumping the log4j2 dependency to version 2.24.3 we
started noticing extra logging output during the build. This
issue has been reported at https://github.com/apache/logging-log4j2/issues/3100
and is due to the change to the AbstractConfiguration class
in https://github.com/apache/logging-log4j2/pull/3043. As
described in log4j2 issue #3100 and at
https://logging.apache.org/log4j/2.x/manual/status-logger.html
the status configuration element is now deprecated and we
are advised to use a system property instead.

Closes #5453