improved validation of load plan json
Converting bulk load plan json to a LoadPlan was not very strict and for
cases that it did fail it threw many different runtime exceptions.  Made
validation of the json more strict and made the code always throw the
same exception type when json is not valid. Added unit test for many
different cases of illegal input.