Use General ScheduledExecutorThreadPool where possible
Noticed several places in the code where different thread
pools were being created based on the GENERAL_THREADPOOL_SIZE
property instead of re-using the one available via the
ServerContext.

TabletServerBatchWriter was doing the same thing. I added
a method on ClientContext similar to the ServerContext method
for getting the shared ScheduledExecutorThreadPool and used
this in TabletServerBatchWriter instead of creating a new one
and shutting it down.