Changes to upgrade framework for 4.0 (#5383)
This change builds on the `accumulo upgrade` utility
added in 2.1.4. See #5438 for more information. This
commit adds the `--start` option to UpgradeUtil which
checks that `accumulo upgrade --prepare` was run on
the instance after it was shut down, performs pre-upgrade
validation, performs any upgrade steps that need to occur
before the Manager is started, and creates a mandatory
maker node in ZooKeeper that enables the Manager
to start.



Co-authored-by: Keith Turner <kturner@apache.org>
Co-authored-by: Daniel Roberts <ddanielr@gmail.com>