fixes intermittent failure in ExternalCompactionMetricsIT (#5475)
ExternalCompactionMetricsIT was looking for some metrics to be greater
than zero when its possible it may never see that.  Relaxed the
test to look for >=0 instead of >0 for these metrics because once the
compactors are started the queues are emptied and after that the values
are always zero.  Earlier in the test there is a place where it seems
like the metrics could reliably be seen w/ a value >0, so added some
checks there.  Earlier in the test compactors are not started and the
compaction queue should have stuff in it causing >0 metrics.

Also added some more debug logging in case the test ever gets stuck in
the first loop.  The debug logging in the second loop was useful in
tracking this problem down.