Added property for setting setDropBehind on compaction input streams (#5473)
In #3079 we modified the FileCompactor to call FSDataInputStream.setDropBehind
on compaction input files. Created a table property to allow the
user to specify which file types should setDropBehind should be used
with (default all). Calling setDropBehind on some files, like Bulk Import
files and files assigned to src or dest tables as part of a clone operation,
could cause an issue as the HDFS blocks for the files may not be in the
page cache on the DataNode machine.