Disabled metrics for ThreadPool used for resending failed mutations
The TabletServerBatchWriter uses a ScheduledExecutorThreadPool for
dealing with failed mutations. The TabletServerBatchWriter code
was calling ThreadPools.createGeneralScheduledExecutorService
which creates the ThreadPoolExecutor with metrics enabled. If more
than one TabletServerBatchWriter is created, then debug messages
start showing up in the logs saying that the associated Gauge's
have already been registered and will be ignored. This change
calls the same underlying code, but disables metrics for the
thread pool