Log user attempting to use Authorization they don't have on server (#5480)
This change removes the logging of a stack trace with an unhelpful
message that was done only for batch scans and replaces it with a
log message that notes the specific Authorization that the user is
trying to use that they don't have. This logging is only done on
the server side, the client receives a BAD_AUTHORIZATIONS error.