Fixes bug with getting disk usage of the ROOT table (#5479)
* Fixes bug with getting disk usage of the ROOT table

Code did not consider ROOT table, and was always returning 0 for the
disk usage as it would scan the ROOT table if the input was the METADATA
table and would scan the METADATA table otherwise. Should have been
scanning ZK if the input was ROOT. Changed to use Ample instead, which
handles this issue nicely.

* rewrote TableDiskUsageTest