Created HostRegexTableLoadBalancer test default pool, fixed bug (#5492)
Users can define a regex for the default pool. The code already supports
this, but we didn't have tests for it. Also fixed a bug in the balance method
with the out of bounds check. The code was using `Map.getOrDefault`
which returns the default value, but doesn't insert it into the map. The
side effect of this is that the out of bounds checks were not being performed
because the condition could never be true. Modified the code to use
`Map.computeIfAbsent` which will insert the value into the map for
subsequent checks.

Closes #5489