adds prune command to accumulo-cluster
As the cluster.yaml file is changed it can leave extra server process
around. The accumulo-cluster script currently does not have a way to
look for extra servers processes, it only operates on what is specified
in the cluster.yaml file.

This commit adds the ability to remove servers that exist in zookeeper
but would not be expected based on the cluster.yaml and current pid
files.

Co-authored-by: Daniel Roberts ddanielr <ddanielr@gmail.com>