Bump Apache Parent POM from version 31 to 34 (#5515)
Bumped parent pom version from 31 to 34. Version 34
of the parent pom includes a change that moves the
inclusion of the Apache snapshop repositories to a
profile that needs to be activated. This can be done
by adding '-Dapache.snapshots=true' to the MAVEN_OPTS
environment variable. Modified the build section of
the README to include these instructions.