Fix race condition in Manager.partitionMigrations
in the Manager to new columns in the root and metadata tables.
The Manager.partitionMigrations method was changed to scan the
root and metadata tables migration column to gather the current
migrations. However, if the root and metadata tables are not
hosted, then this scan will hang until the tablet locations
can be resolved.

ScanServerUpgrade11to12TestIT.testScanRefTableCreation has been
failing since #5416 was merged. The test deletes the scanref
table, shuts down the TabletServers and Manager, then restarts
them. This leaves the root tablet in a state where it needs to
perform recovery. The Manager.partitionMigrations method is
called via the StatusThread, and it appears that the Manager
starts up and the StatusThread gets hung trying to scan the
root tablet (it's waiting for a location). Meanwhile, the root
tablet can't be assigned a location because the Manager.tserverStatus
map is not populated, which is done from the StatusThread as well.

Also modified the IT to set the filesystem to the RawLocalFileSystem
so that warnings about missing checksum files were not in the logs.