Avoids creating a batch writer per completed ext compaction
The compaction coordinator was creating a batch writer per a completed
external compaction.  The batch writer was used to write a single small
mutation.  This caused a lot of thread creation and RPCs in the
coordinator.  Changed the coordinator to use a single batch writer for
these mutations.  Also update some logging in the coordinator to lower
levels to trace and add some timing information.

Co-authored-by: Daniel Roberts ddanielr <ddanielr@gmail.com>