Shared mini test suite (#5536)
Created SimpleSharedMacTestSuite which starts Mini, runs all of the tests with
the SIMPLE_MINI_CLUSTER_SUITE tag, then shuts down Mini. Annotated most of
the tests that extended SharedMiniClusterBase and called startMiniCluster with
no configuration callback with the SIMPLE_MINI_CLUSTER_SUITE tag. These tests
with that tag did not use a configuration callback where most likely to not do
destructive things with Mini. ManagerApiIT was one exception and I did not put
the tag on this class.

Modified SharedMiniClusterBase so that stopMiniCluster would not stop the
cluster if STOP_DISABLED was set to true, which it is in the new test suite class.
This enables a single test class to be run and behave normally, but also allows
the class to be part of the test suite and not shut down Mini after the tests have
run. Modified stopMiniCluster to instead delete all user tables and namespaces
when STOP_DISABLED is set to true to clean up tables from the test class. Leaving
the tables around in the instance worked for most tests, but not for tests that
counted tables and expected a certain number (CreateTableIT) or for tests that
kicked off a compaction and waited for it to finish (ComprehensiveIT.testCompaction)
as the single Compactor may have been busy compacting other tables.

Modified the pom to have two failsafe executions, one for the ITs without the
SIMPLE_MINI_CLUSTER_SUITE and one for the test suite.

Closes #5529