Added SharedBatchWriterQueue for SharedBatchWriter object
Added SharedBatchWriterQueue interface for use with the
SharedBatchWriter. This allows producers to put work on
the queue while the single consumer thread writes to the
configured table.

Modified the CompactionFinalizer's use of the SharedBatchWriter
to create a SharedBatchWriterQueue for the External Compaction
Final State mutations and wired up the DeadCompactionDetector
to inspect the SharedBatchWriterQueues for queued writes when
determining whether a compaction is dead or not.