makes balancing a noop when balancer setup fails. (#5530)
* WIP makes balancing a noop when balancer setup fails.

In the case where setting up a balancer fails the manager currently
falls back to using the SimpleLoadBalancer.  This fallback behavior
could be very unhelpful and one problem (bad balancer config) into
multiple problems as the simple load balancer may start assigning
tablets in a way that is detrimental.

The reason this is a draft is to get feedback on the approach. New test
are needed, but those will be a lot more work than this simple change.
Did not want to write test if we settle on a different approach. Need to
test the following for system balancer and per table balancers.

 * Configuring a balancer class name that does not exists
 * Configuring a balancer class that is not of the correct type
 * Configuring a balancer class that throws an exception in it
   constructor
 * Configuring a balancer class that throws an exception when init is
   called

* Rename to DoNothingBalancer. Updates log messages

Renames the Derelict Balancer to the DoNothing Balancer.
Updates the log messages to show that the ignore message is due to a
failed state.

* fix formatting

* added test and fixed bugs

* format code

* code review updates

* code review update

---------

Co-authored-by: Daniel Roberts ddanielr <ddanielr@gmail.com>