fixes external compaction race condition (#5570)
* fixes external compaction race condition

In there tablet server there was a race condition between the set of
online tablets and set of external compactions.  When a tablet loaded w/
existing external compaction it would add to the external compaction set
first and the online tablet set second. This could cause cause code that
cleans up external compactions to remove something it saw in  one set
and not the other.  Considered opening tablets to avoid this race.  Also
adjusted rpc code to avoid the race.

* fix race condition in fix and improve logging

* improve logging

* avoid copying immutable set while lock is held