Fixed issue with balancer property updates
The change in #5530 added code to clear the
property cache in initializeBalancer because
the property change was not immediately seen.
BrokenBalancerIT was changing the property
then checking the balance to see if the new
balancer had taken effect.

This commit changes BrokenBalancerIT to give
it more time in the test thread for the
property change to take effect in the Manager
and changes how the Manager responds to system
property changes so that the property cache
invalidation can be removed.

Related to #5530