fixes chaotic balancer IT
When assigning the root tablet after other tablets were assigned, the
getAssignment function in ChaoticLoadBalancer would never return an
assignment. This would cause the test to hang until it timed out.  Not
sure what the intention of the code was, but made it do something
different that would always eventually yield an assignment.