Created bind addr property, add help for -o options
This change includes a new property, rpc.bind.addr, which
was included in 3.0 as general.process.bind.addr in #3192,
backports ConfigOptsTest from #3192, and changes to
AbstractServer to use either the '-a' or '-o rpc.bind.addr'
bind address as the hostname.

This change also includes changes to ConfigOpts such that
when the '-h' option is used **and** the command is a
server name, then it prints out possible property overrides
to use with the '-o' option.

For example, 'accumulo rfile-info -h' and 'accumulo tserver -x'
just print the normal usage information. However, 'accumulo tserver -?'
or 'accumulo tserver -h' will print an expanded usage information
that includes properties that the user can override on the
command line for that server.

Related to #4976
Closes #5430