fixes incorrect file range clipping in deleterows (#5590)
The delete rows table operation was clipping a file to a row range like
(f,f] and adding this to tablet with a row range of (f,g].  The files row
range did not overlap the tablets row range causing reads of the tablet
to fail with the changes in #5340.  This was causing DeleteRowsIT to
fail, modified the code to avoid adding the non overlapping row range
and now DeleteRowsIT passes.

Made a lot of logging changes while tracking this down, kept those in
this commit.