Added property for the advertise address
This is a follow-on to #5599 to add a separate property
for the server process advertise address. The bind address
should be used for creating the Thrift server, and the
advertise address should be used in all other locations (ZK,
metadata locations, etc.). The advertise address, if not
speicified, should default to the bind address.

Related to #5599