optimizes tablet metadata filter
TabletMetadataFilter had a set of columns needed for filtering available
to it but did not use them.  Modified it to use the columns during
filtering.  The comment in the change provides more details.

Attempted to modify RowFilter to support this, but the way it uses its
deep copy is inverted from what is needed for this case.