Bound seek range using families in iter stack
When using scanner or batch scanner and setting a range+column on the
scanner Range.bound will be called by the scanner to narrow the range.
However when seeking in an iterator on the server side, nothing similar
is done.  Added this narrowing to ColumnFamilySkippingIterator which
will cause any server side iterator seeks to be narrowed using
Range.bound().

Manually did some test to ensure nothing in the code was currently doing
this.  Added some prints to the rfile code and had server side iterators
seek with a range over an entire row w/ some column families.  Did not
see any narrowing happen w/o this change.  Wanted to make sure it was
not being done by code somewhere in the iter stack that did not call
Range.bound().