Looks ahead more than once in compaction planning. (#5588)
When planning compactions for a tablet that currently has compactions
running, then planning process will estimate what the size of the file
the running compactions would produce and see if its more optimal to
wait for these.  When doing this it only looks ahead once.  This change
makes the planning process look ahead more for files to compact.