compresses migrations used in metadata table scans
The manager sends the set of migrations to each metadata tablet server
when it scans the metadata table. For a large set of migrations and many
metadata tablet servers the cost of repeatedly sending the same data
adds up. This change compresses the migrations once and then each rpc
to a tserver sends the compressed data.  Decompression will happen in
parallel on the tservers.  The one time cost of compression should
eventually be worthwhile because of the reduced amount of data sent from
manager to many tservers.

A much better way to solve this overall problem is #5416, but that would
be a very disruptive to change to make in 2.1. This change and #5626
can improve throughput for this situation in 2.1 w/ minor changes.