fixes bug in seeking iterator w/ disjoint range and columns
In #5621 a bug was introduced where if an iterator seeks with range and
columns families that are disjoint it will throw an exception.  This was
caused by Range.bound() that had this behavior.

To fix this copied the code in Range.bound() into RangeImpl and made a
very slight modification.  This was done to preserve the existing
behavior of the public API method Range.bound() and reuse 99% of its
code and test for a slightly different behavior.  Also documented the
behavior of Range.bound(), considered changing its behavior but this
subtle change in runtime behavior seemed dicey w/ hard to predict
outcomes (like it would change the runtime behavior of Scanner and
BatchScanner which currently use Range.bound()).

Now when the range and column families are disjoint
ColumnFamilySkippingIterator will seek its underlying iterator with a
new empty range.

This bug was found by BigRootTabletIT.  After this change that test
passes.