Upgraded spotbugs-maven-plugin to 4.9.3.0 (#5623)
Upgraded plugin which required changes to existing
classes as spotbugs was complaining about encapsulation
issues. Modified existing configuration to ignore
two new bug detectors, ConstructorThrow and
SharedVariableAtomicityDetector. Fixing the code
to enable the ConstructorThrow detector would require
a lot of refactoring of objects that throw exceptions
from their constructor. The SharedVariableAtomicityDetector
was complaining about a lot of non-atomic variable manipulation
and access to variables from different threads that might
require synchronization. Each occurrence likely needs to
evaluated to determine if it's an issue or not. For example,
it was suggesting that hasTop needs to be marked as
volatile in an iterator class because it's set and read
in different methods. However, iterators don't support
multi-threaded access.

Removed the spotbugs suppression to CachableBlockFile
that was added in #5381.

Closes #5382