Modified TableMapping.update to only check root and meta exist
TableMapping.update was added as part of #5451 and it validated
that the accumulo namespace contained all of the system tables.
This is technically correct, except when upgrading. Since this
is used as part of the client, there is not a good way to
determine whether or not the instance is upgrading to implement
an alternate check. This change modifies the check in
TableMapping.update to only check that root and meta tables
exist to allow the upgrade code to run.

Closes #5640