Modified TabletsMetadata to run filters on root tablet
Filters on TabletMetadata were not being applied to the root tablet
metadata causing an exception in in Manager.shouldCleanupMigration.
The code was applying a filter to only return tablet metadata that
contained a migration, then checking that the migration was not
null. Because the filter was not being applied to the root tablet,
it was being returned as matching the filter but with no migration.

Applying the fix in TabletsMetadata to run the filter on the root
tablet metadata caused a subsequent issue in the GcWalsFilter where
acceptTablet() would return null. Modified TabletMetadataFilter
to protect against this condition.

Closes #5643