BASH completion script for accumulo commands
Created a BASH completion script in the assemble/bin folder called _accumulo_completions that allows the user to complete the accumulo, accumulo admin, and accumulo ec-admin commands to streamline usage and testing of Accumulo. Gives the user suggestions of what they can enter next given what's already entered on the command line. Created a README file in the assemble folder that describes the setup and usage of the script.

Closes #4808