Send error back from TabletManagementIterator on invalid metadata
Modified the TabletManagementIterator to invoke getPrevEndRow()
on the constructed TabletMetadata object to catch a the raised
IllegalStateException and return it to the Manager. The existing
try/catch around computeTabletManagementActions would likely
throw the IllegalStateException when tm.getExtent is called
when trying to log an error, preventing the error from being
retured to the Manager.

There is an existing test case, TabletMetadataTest.testAbsentPrevRow,
which demonstrates that an IllegalStateException is raised when
TabletMetadata.convertRow is called on a set of keys that don't
contain a prevEndRow.

Closes #5658