Moves balancing code out of Manager class (#5537)
The balancing code needs some improvements.  Currently the code is
entangled in the manager code in odd places (like under some of the
balancing code is un the Manager.StatusThread class).  This change moves
the balancing code to its own class w/o making any changes to the code
(except fixing two small existing bugs in shouldCleanupMigration()).

Making this change as a first step in making the balancer more
independent of other manager code and threads.

Co-authored-by: Dom G. <domgarguilo@apache.org>