Created new static methods that easily create TabletId objects (#5678)
* Created new static methods that easily create TabletId objects

Created the static overloaded of() methods in TabletId.java that creates TabletId objects using TabletIdImpl. Implemented testing in TabletIdTest.java that ensure TabletId is created correctly and it's fields can be properly accessed and used. This is being submitted as an update to include new methods in 2.1. Changes will be made later in a separate PR to target version 4.0.0 in order to replace existing Mock tests of TabletId with these methods.

Closes #5666

---------

Co-authored-by: Kevin Rathbun <kevinrr888@gmail.com>