Allows limiting per tablet information retrieved
When calling TableOperations.getTabletInformation() it currently
retrieves a lot of tablet metadata.  If this information is not of
interest then a lot of time could be spent gathering information that is
never used (especially for the per tablet file data).  Modified
TableOperations.getTabletInformation() to allow specifying a subset of
tablet metadata to fetch.

This is a WIP, it still needs test.  Posting API changes fore review
before adding test.