Allow IT classes to override server class at the resource group level (#5692)
Modified MiniAccumuloConfigImpl.setServerClass such that the user
can supply an override class to use per resource group. If no class
override is specified, then the default implementations are used.
When overriding classes for the Compactor in an IT, the default
resource group may need to use the default implementation to
ensure that the root and metadata tables are compacting.

Closes #5683