fixes bug where external compaction never starts
Ran into an issue when running random walk bulk test where a compaction
that needed to happen was never running.  The code fixed in this PR used
to change the state to running and then forget about it in some cases.
However [this code][1] would still remember it.  That would cause this
[this code][2] to filter out planned jobs when there were running
compactions in submitted jobs. Hopefully this change will cause [this
code][3] to filter this from submitted jobs.

[1]:https://github.com/apache/accumulo/blob/5bba2f3cf36717d0e2fabbf4abb93c3455c0619e/server/tserver/src/main/java/org/apache/accumulo/tserver/compactions/CompactionService.java#L84
[2]:https://github.com/apache/accumulo/blob/5bba2f3cf36717d0e2fabbf4abb93c3455c0619e/server/tserver/src/main/java/org/apache/accumulo/tserver/compactions/CompactionService.java#L186
[3]:https://github.com/apache/accumulo/blob/5bba2f3cf36717d0e2fabbf4abb93c3455c0619e/server/tserver/src/main/java/org/apache/accumulo/tserver/compactions/CompactionService.java#L359-L365