Add .mvn/maven.config file for snapshots repo
Add .mvn/maven.config file for local maven configuration that activates
the use-apache-snapshots profile from the Apache parent POM, so that we
can resolve the accumulo-access-1.0.0-SNAPSHOT dependency during
development without developers needing to set up any extra local
configuration.

This should be removed when the snapshot repository is no longer needed
to resolve SNAPSHOT dependencies.

This is related to #5515