Cause Compactor to exit when error loading classes
AccumuloVFSClassLoader.getContextClassLoader throws an
UncheckedIOException when there is an issue getting the
ClassLoader for a context name. This runtime exception
escapes all the way up to the calling code in Compactor
and TabletServer, which fails the compaction and moves
on to the next compaction. If there is an issue getting
the ClassLoader for the context once, then it's likely
to happen again. It's probably not safe to terminate
the TabletServer in this case, but is likely safe for
the Compactor.

This change captures the RuntimeException in the
FileCompactor.compactLocalityGroup where the iterator
stack is created and raises a new checked exception
which is handled in the calling code.