Avoid client side errors during manager shutdown. (#5725)
Client side table operations that used fate could see exceptions during
manager shutdown before this fix.  After this fix hopefully that is not
the case.  Was occasionally seeing this problems in a ShutdownIT test.
Modified ShutdownIT to more reliably cause the problem before this fix,
after this fix the test is passing consistently.