avoids reassignment of hosted tablet during shutdown
When a tablet server is gracefully shutdown it notifies the manager it
is shutting down.  The manager would remove the server from the set of
live tservers which caused the manger to treat all tablets on that
tserver as assigned to dead tservers.

Saw this when debugging a failure with GracefulShutdownIT.  Saw the
manager initiate log recovery and reassignment for the root tablet
before it unloaded from the tserver that was shutting down.  This caused
errors with the metadata tablet consistency check during unload of the
root tablet because the manager had modified the tablet metadata before
unload.

This fix removes the notification to the manager and replaces it with a
tablet server check to ignore loading tablets during tserver shutdown.
Created the new OpeningTablets class because the two sets it contains
are very tightly coupled, so seemed best to create a new class rather
than add those two variables to TabletServer class.