Avoids reading all fate txs from zookeeper when reserving (#5728)
Each fate thread that was looking for a single fate operation to work on
would read all fate txs from zookeeper.  This does N^2 work and would
place a lot of load on zookeeper and the manager as the number of total
fate transaction increased.  This change avoids reading everything from
zookeeper when looking for a single thing to work on.

Ran the random walk bulk test w/ this change and saw a much lower CPU
usage for the zookeeper server.  Also ran `mvn verify -Psunny`
successfully against these changes.

fixes #5712