Allows the monitor to be hosted behind a proxy (#5729)
Changes the templates to use a base href path vs hardcoded absolute
paths and changes links in js files to use the base href value.

Allows the user to set a system property to define the root context of
the monitor.

Update the js files and functions created by FreeMarker to use the
contextPath when making requests.

Adds test for EmbeddedWebServer to check if the contextPath is correct.
Also adds preconditions to Monitor.java to check the format of the root
context value.